+++ app/src/main/java/com/baumannsw/lewind/windData/StationDownloaderCaller.java
@@ -1,6 +1,6 @@
 package com.baumannsw.lewind.windData;
 
 public interface StationDownloaderCaller {
-    void onStationDownloadCompleted(StationData data, int id);
-    void onStationDownloadFailed(String errorMessage, int id);
+    void onStationDownloadCompleted(StationData data, long id);
+    void onStationDownloadFailed(String errorMessage, long id);
 }
